chore(distrobox): enter distrobox then zellij automatically
I use Fedora Atomic desktops which discourage the installation of
developer tools to the base operating system. Instead, they recommend
containerising development or running dev in a distrobox. I've found
developing inside a single Debian distrobox to be easier.

This change makes it so that when I start GhosTTY, the dev distrobox is
entered, then the Zellij launchpad session is started and created if it
doesn't exist.

This means that I'm seamlessly inside of a mutable dev environment when
I open GhosTTY.