perf: improve performance of rust-analyzer
rust-analyzer and rustc were making my laptop slow to a crawl. This
change limits the system resources it can use (slowing it down), and enables features that improve its performance (speeding it up).

End result is that I should be able to use my laptop during the initial
scan when opening a Rust project.